Se agregó el mapeo de los identifiers de Person a Patient.

+++ src/main/java/org/saintmartinhospital/fhir/service/patient/PatientService.java
@@ -4,6 +4,6 @@ import org.hl7.fhir.r4.model.Patient;
 
 public interface PatientService {
 
-	Patient findById( Long patientId );	
+	Patient findById( Integer patientId );	
 	
 }
